Move process sequencing classes into workflow package
