+++ src/main/java/demo/Demo.java
@@ -14,9 +14,9 @@ import org.opensearch.client.Client;
 import org.opensearch.client.node.NodeClient;
 import org.opensearch.common.SuppressForbidden;
 import org.opensearch.common.io.PathUtils;
-import org.opensearch.flowframework.template.ProcessNode;
 import org.opensearch.flowframework.template.Template;
-import org.opensearch.flowframework.template.WorkflowProcessSorter;
+import org.opensearch.flowframework.workflow.ProcessNode;
+import org.opensearch.flowframework.workflow.WorkflowProcessSorter;
 import org.opensearch.flowframework.workflow.WorkflowStepFactory;
 
 import java.io.IOException;
