+++ src/main/java/org/opensearch/flowframework/FlowFrameworkPlugin.java
@@ -16,7 +16,7 @@ import org.opensearch.core.common.io.stream.NamedWriteableRegistry;
 import org.opensearch.core.xcontent.NamedXContentRegistry;
 import org.opensearch.env.Environment;
 import org.opensearch.env.NodeEnvironment;
-import org.opensearch.flowframework.template.WorkflowProcessSorter;
+import org.opensearch.flowframework.workflow.WorkflowProcessSorter;
 import org.opensearch.flowframework.workflow.WorkflowStepFactory;
 import org.opensearch.plugins.Plugin;
 import org.opensearch.repositories.RepositoriesService;
