+++ src/main/java/org/opensearch/flowframework/workflow/WorkflowProcessSorter.java
@@ -6,13 +6,13 @@
  * this file be licensed under the Apache-2.0 license or a
  * compatible open source license.
  */
-package org.opensearch.flowframework.template;
+package org.opensearch.flowframework.workflow;
 
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
-import org.opensearch.flowframework.workflow.WorkflowData;
-import org.opensearch.flowframework.workflow.WorkflowStep;
-import org.opensearch.flowframework.workflow.WorkflowStepFactory;
+import org.opensearch.flowframework.template.Workflow;
+import org.opensearch.flowframework.template.WorkflowEdge;
+import org.opensearch.flowframework.template.WorkflowNode;
 
 import java.util.ArrayDeque;
 import java.util.ArrayList;
